WIIS-9444: carrier logo error message

+++ src/Controller/TransporteurController.php
@@ -132,7 +132,7 @@ class TransporteurController extends AbstractController
                 $entityManager->remove($attachmentToRemove);
             }
             $transporteur->addAttachment($logo);
-        } else if ($carrierId && !$transporteur->getAttachments()->isEmpty()) {
+        } else if ($carrierId && !$transporteur->getAttachments()->isEmpty() && !$data['keep-logo']) {
             $attachmentToRemove = $transporteur->getAttachments()[0];
             $transporteur->removeAttachment($attachmentToRemove);
             $entityManager->remove($attachmentToRemove);
